Update index.html
This update adds a new graph that shows how often White and Black are played. It also includes a methodology statement at the bottom, while adjusting the original 2.5% hardcoded boost for Black (because it had such a wide margin of games less than white) and changes it to a dynamic range from 0 - 2.5% for each side. A "last updated" section is also new.