Update index.html
Added a new graph that shows how often White and Black are played on each side. Added a dynamic range from 0% - 2.5% if either White or Black have more games than each other instead of a hardcoded boost for Black (because it had way less games for a long period of time). Added a new methodology statement at the bottom, and added a "last updated" section.